refactor semaphore and address feedback
