+++ c/datasetService.c
@@ -37,7 +37,9 @@
 #include "httpserver.h"
 #include "dataservice.h"
 #include "json.h"
+#include "envService.h"
 #include "datasetjson.h"
+#include "datasetlock.h"
 #include "logging.h"
 #include "zssLogging.h"
 
@@ -148,7 +150,7 @@ static int serveDatasetEnqueue(HttpService *service, HttpResponse *response){
     char *percentDecoded = cleanURLParamValue(response->slh, l1);
     char *filenamep1 = stringConcatenate(response->slh, "//'", percentDecoded);
     char *filename = stringConcatenate(response->slh, filenamep1, "'");
-    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_DEBUG, "Taking enqueue on: %s\n", filename);
+    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_DEBUG, "Taking enqueue on: %s username: %s\n", filename, request->username);
     respondWithEnqueue(response, filename, TRUE);
   }
   else if (!strcmp(request->method, methodDELETE)) {
@@ -156,10 +158,8 @@ static int serveDatasetEnqueue(HttpService *service, HttpResponse *response){
     char *percentDecoded = cleanURLParamValue(response->slh, l1);
     char *filenamep1 = stringConcatenate(response->slh, "//'", percentDecoded);
     char *filename = stringConcatenate(response->slh, filenamep1, "'");
-    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_DEBUG, "Releasing enqueue on: %s\n", filename);
-    
+    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_DEBUG, "Releasing enqueue on: %s , username: %s \n", filename, request->username);
     respondWithDequeue(response, filename, TRUE);
-    
   }
   else {
     jsonPrinter *out = respondWithJsonPrinter(response);
@@ -176,8 +176,6 @@ static int serveDatasetEnqueue(HttpService *service, HttpResponse *response){
 
     finishResponse(response);
   }
-  zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_DEBUG2, "end %s\n", __FUNCTION__);
-  zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_DEBUG, "Returning from %s\n", __FUNCTION__);
   return 0;
 }
 
@@ -289,6 +287,51 @@ void installDatasetMetadataService(HttpServer *server) {
                     makeStringParamSpec("includeUnprintable", SERVICE_ARG_OPTIONAL,
                       makeIntParamSpec("workAreaSize", SERVICE_ARG_OPTIONAL, 0,0,0,0, NULL))))))))));
   registerHttpService(server, httpService);
+}
+
+static void readDatasetSettings(int* heartbeat, int* expiry) {
+  JsonObject *datasetLockSettings = readEnvSettings("ZSS_DATASET");
+  if (datasetLockSettings == NULL) {
+    *heartbeat = 0;
+    *expiry = 0;
+  }
+  *heartbeat = jsonObjectGetNumber(datasetLockSettings, "ZSS_DATASET_HEARTBEAT");
+  zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_INFO,"heartbeat_loop_time %d\n", *heartbeat);
+  *expiry = jsonObjectGetNumber(datasetLockSettings, "ZSS_DATASET_EXPIRY");
+  zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_INFO,"heartbeat_expiry_time %d\n", *expiry);
+}
+
+static int serveDatasetHeartbeat(HttpService *service, HttpResponse *response){
+  zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_INFO, "begin %s\n", __FUNCTION__); 
+  HttpRequest *request = response->request;
+  if (!strcmp(request->method, methodPOST)) { 
+    resetTimeInHbt(request->username);
+    respondWithMessage(response, HTTP_STATUS_NO_CONTENT, "");
+  } else {
+    addStringHeader(response, "Allow", "POST");
+    respondWithError(response, HTTP_STATUS_METHOD_NOT_FOUND, "Method Not Allowed");
+  }
+  return 0;
+}
+
+
+void installDatasetHeartbeatService(HttpServer *server) {
+  zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_INFO, ZSS_LOG_INSTALL_MSG, "dataset Heartbeat");
+
+  HttpService *httpService = makeGeneratedService("datasetHeartbeart", "/datasetHeartbeat/**");
+  httpService->authType = SERVICE_AUTH_NATIVE_WITH_SESSION_TOKEN;
+  httpService->runInSubtask = TRUE;  
+  httpService->doImpersonation = FALSE;
+  httpService->serviceFunction = serveDatasetHeartbeat;
+  registerHttpService(server, httpService);
+
+  int heartbeat;
+  int expiry;
+  readDatasetSettings(&heartbeat, &expiry);
+  //initialize lock tables
+  initLockResources(heartbeat, expiry);
+  // register background handler
+  addZssBackgroudTask(&heartbeatBackgroundHandler,"DATASET_HEARTBEAT_TASK", lockService->heartbeat);
 }
 
 #endif /* __ZOWE_OS_ZOS */
