--- h/datasetHeartbeatService.h
@@ -1,30 +1,0 @@
-
-
-/*
-  This program and the accompanying materials are
-  made available under the terms of the Eclipse Public License v2.0 which accompanies
-  this distribution, and is available at https://www.eclipse.org/legal/epl-v20.html
-  
-  SPDX-License-Identifier: EPL-2.0
-  
-  Copyright Contributors to the Zowe Project.
-*/
-
-#ifndef __DATASET_HEARTBEAT_SERVICE_H__
-#define __DATASET_HEARTBEAT_SERVICE_H__
-
-void installDatasetHeartbeatService(HttpServer *server);
-
-#endif /* __DATASET_HEARTBEAT_SERVICE_H__ */
-
-
-/*
-  This program and the accompanying materials are
-  made available under the terms of the Eclipse Public License v2.0 which accompanies
-  this distribution, and is available at https://www.eclipse.org/legal/epl-v20.html
-  
-  SPDX-License-Identifier: EPL-2.0
-  
-  Copyright Contributors to the Zowe Project.
-*/
-
